fiddle/backend/pyproject.toml
[tool.poetry]
name = "fiddle"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"
packages = [
    { include = "app", from = "." }
]


[tool.poetry.dependencies]
python = "^3.9"
baml = "^0.17.1"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
